Version 2.1
You might see version 2.0 in the commits but it was a mistake that I've made! 
The latest version is 2.1.

Version 2.1: Minor Bug Fixes. (There was also a huge mistake from me while uploading version 2.0 which led to corruption of v2.0 uploaded about an hour ago. But we have v2.1 now!)
Version 2.0: Encrypt any file with any file format or extension.